test(e2e): apply review feedback - dedupe, dejank, kill skipped tests
Following an internal review of the new test suite. Five must-fixes
plus nine nice-to-fix items.

Must-fix:
- Delete TestIssuesUpdate_ClearAssignee — perpetually-skipped tests are
  noise; the missing-feature note belongs in source, not in the suite.
- Seed e2e/testdata/issue_get.golden.json and
  e2e/testdata/mcp_tools_readonly.golden.json so the golden tests run
  on every invocation. Verified stable across Redmine 4.2/5.1/6.1.
- Constrain TestUsers_List/pagination disjoint and TestUsers_ListLocked
  via --name <prefix> so they no longer race against any concurrent
  user activity on the server.
- Pin TestVersions_LockedRejectsAssignment to fail when Redmine accepts
  the assignment (was Skip-on-pass — the test could never fail).
- Merge mcpExtAssertGoldenJSON into the shared assertGoldenJSON helper.

Nice-to-fix:
- Strengthen metadata_test.go: trackers test now pins the CSV header
  shape (CLI-specific value); statuses test keeps the open/closed gate.
- Move TestSearch_Wiki, TestSearch_AllScopes, TestSearch_NoResults into
  search_test.go; TestAPI_NotFoundResponse, TestAPI_Delete,
  TestAPI_FormFlag into api_test.go; TestErrors_ConnectionFailure into
  errors_test.go; TestVersions_LockedRejectsAssignment and
  TestVersions_GetNonexistent into versions_test.go. Delete
  extras_test.go and the duplicate TestErrors_ValidationFailure (kept
  the issues_edge_test.go variant).
- Tighten TestIssuesCreate_EmptySubjectValidationError accept-list to
  the verified validation_failed code.
- Move requireErrorEnvelopeMessage into helpers_test.go so all topical
  files share one helper.
- Pass an explicit anchor through summaryDateWindow + the entry-date
  computations in time_summary_test.go to remove the midnight-UTC race
  between entry creation and the summary window.
- Expose defaultRunnerProfile constant from runner_test.go so
  auth_lifecycle_test.go does not hardcode 'local-e2e'.
- Split TestOutput_NoColor: keep the NO_COLOR assertion as
  TestOutput_NoColorEmitsPlain, drop the FORCE_COLOR half (renderer
  ignores it without a TTY, so it tested nothing).
- Collapse the 6 completion tests into 2 parameterized tests.

Verified: 88 PASS / 0 FAIL / 0 SKIP across Redmine 4.2/5.1/6.1.
