+++ e2e/completion_test.go
@@ -8,88 +8,54 @@ import (
 	"testing"
 )
 
-// The completion command is built from cobra metadata and never contacts the
-// Redmine server, but we still gate behind requireE2E for consistency with the
-// rest of the suite (and so the binary build in TestMain is amortized across
-// all tests).
-
-// TestCompletion_BashGenerates verifies that `redmine completion bash`
-// produces a non-empty bash completion script with the cobra bash shape
-// and a per-subcommand handler. Cobra registers the completion via
-// `complete -o default -F __start_redmine redmine`, so __start_redmine is
-// the stable marker.
-func TestCompletion_BashGenerates(t *testing.T) {
+// TestCompletion_Generates exercises every shell that the completion command
+// supports. The cobra-generated output never contacts Redmine but we keep the
+// tests under the e2e build tag so TestMain's binary build is amortised.
+//
+// For each shell we assert one structural marker (proves cobra ran for that
+// shell) plus, where the output is statically templated, a known subcommand
+// name. zsh's completion is dynamic and does not embed subcommands, so we
+// only assert the static markers there.
+func TestCompletion_Generates(t *testing.T) {
 	requireE2E(t)
 	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
 
-	stdout := r.run(t, "completion", "bash")
-	assertCompletionScript(t, "bash", stdout,
-		"__start_redmine", "_redmine_issues", "_redmine_projects")
-}
-
-// TestCompletion_ZshGenerates verifies that `redmine completion zsh` produces
-// a non-empty zsh completion script with the expected zsh shape. Note: cobra's
-// zsh completion is dynamic — it does NOT bake subcommand names into the
-// script, so we only assert on the static markers.
-func TestCompletion_ZshGenerates(t *testing.T) {
-	requireE2E(t)
-	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
-
-	stdout := r.run(t, "completion", "zsh")
-	assertCompletionScript(t, "zsh", stdout, "#compdef redmine", "_redmine")
-}
-
-// TestCompletion_FishGenerates verifies that `redmine completion fish`
-// produces a non-empty fish completion script with the expected fish shape.
-func TestCompletion_FishGenerates(t *testing.T) {
-	requireE2E(t)
-	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
-
-	stdout := r.run(t, "completion", "fish")
-	assertCompletionScript(t, "fish", stdout, "complete -c redmine", "issues")
-}
-
-// TestCompletion_PowerShellGenerates verifies that `redmine completion
-// powershell` produces a non-empty PowerShell completion script.
-func TestCompletion_PowerShellGenerates(t *testing.T) {
-	requireE2E(t)
-	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
-
-	stdout := r.run(t, "completion", "powershell")
-	assertCompletionScript(t, "powershell", stdout, "Register-ArgumentCompleter", "redmine")
-}
-
-// TestCompletion_InvalidShell asserts that cobra rejects an unknown shell with
-// a non-zero exit. The completion command uses cobra.OnlyValidArgs so any
-// value outside {bash,zsh,fish,powershell} must fail before script generation.
-func TestCompletion_InvalidShell(t *testing.T) {
-	requireE2E(t)
-	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
-
-	r.runExpectError(t, "completion", "not-a-shell")
-}
-
-// TestCompletion_NoArgs asserts that omitting the shell argument fails the
-// cobra.ExactArgs(1) gate.
-func TestCompletion_NoArgs(t *testing.T) {
-	requireE2E(t)
-	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
-
-	r.runExpectError(t, "completion")
-}
-
-// assertCompletionScript fails the test if stdout is empty or any required
-// marker is missing.
-func assertCompletionScript(t *testing.T, shell string, stdout []byte, markers ...string) {
-	t.Helper()
-	if len(bytes.TrimSpace(stdout)) == 0 {
-		t.Fatalf("completion %s: stdout is empty", shell)
+	cases := []struct {
+		shell   string
+		markers []string
+	}{
+		{"bash", []string{"__start_redmine", "_redmine_issues", "_redmine_projects"}},
+		{"zsh", []string{"#compdef redmine", "_redmine"}},
+		{"fish", []string{"complete -c redmine", "issues"}},
+		{"powershell", []string{"Register-ArgumentCompleter", "redmine"}},
 	}
-	got := string(stdout)
-	for _, marker := range markers {
-		if !strings.Contains(got, marker) {
-			t.Fatalf("completion %s: stdout missing %q\nstdout:\n%s",
-				shell, marker, got)
-		}
+	for _, tc := range cases {
+		t.Run(tc.shell, func(t *testing.T) {
+			stdout := r.run(t, "completion", tc.shell)
+			if len(bytes.TrimSpace(stdout)) == 0 {
+				t.Fatalf("completion %s: stdout is empty", tc.shell)
+			}
+			got := string(stdout)
+			for _, marker := range tc.markers {
+				if !strings.Contains(got, marker) {
+					t.Errorf("completion %s: stdout missing %q\nstdout:\n%s",
+						tc.shell, marker, got)
+				}
+			}
+		})
 	}
+}
+
+// TestCompletion_Rejects covers the cobra arg-validation paths: an unknown
+// shell name and a missing arg both exit non-zero before script generation.
+func TestCompletion_Rejects(t *testing.T) {
+	requireE2E(t)
+	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
+
+	t.Run("invalid shell", func(t *testing.T) {
+		r.runExpectError(t, "completion", "not-a-shell")
+	})
+	t.Run("no args", func(t *testing.T) {
+		r.runExpectError(t, "completion")
+	})
 }
