+++ e2e/metadata_test.go
@@ -2,34 +2,39 @@
 
 package e2e
 
-import "testing"
+import (
+	"bytes"
+	"encoding/csv"
+	"slices"
+	"testing"
+)
 
-// TestMetadata_TrackersList verifies that the read-only `trackers list`
-// command returns a non-empty collection with usable id/name fields.
-func TestMetadata_TrackersList(t *testing.T) {
+// TestMetadata_TrackersListCSV verifies the CSV output shape of `trackers
+// list` since the existing `firstTrackerName` helper already covers the JSON
+// path. Pinning the headers protects renames in the CSV column layout.
+func TestMetadata_TrackersListCSV(t *testing.T) {
 	requireE2E(t)
 	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
 
-	var trackers []struct {
-		ID   int    `json:"id"`
-		Name string `json:"name"`
+	stdout := r.run(t, "trackers", "list", "--output", "csv")
+	rows, err := csv.NewReader(bytes.NewReader(stdout)).ReadAll()
+	if err != nil {
+		t.Fatalf("decode trackers CSV: %v\nstdout:\n%s", err, stdout)
 	}
-	r.runJSON(t, &trackers, "trackers", "list")
-	if len(trackers) == 0 {
-		t.Fatal("trackers list returned no trackers")
+	if len(rows) < 2 {
+		t.Fatalf("expected header + at least 1 tracker row, got %d rows\nstdout:\n%s", len(rows), stdout)
 	}
-	for _, tr := range trackers {
-		if tr.ID == 0 {
-			t.Errorf("tracker has zero ID: %+v", tr)
-		}
-		if tr.Name == "" {
-			t.Errorf("tracker has empty name: %+v", tr)
-		}
+	wantHeaders := []string{"ID", "Name", "Description"}
+	if !slices.Equal(rows[0], wantHeaders) {
+		t.Fatalf("trackers CSV header mismatch:\n got: %v\nwant: %v", rows[0], wantHeaders)
 	}
 }
 
-// TestMetadata_StatusesList verifies that `statuses list` returns the default
-// issue statuses with both open and closed entries.
+// TestMetadata_StatusesList verifies the JSON shape of `statuses list` and
+// that both an open and a closed status are present, mirroring the open/closed
+// gating that issues_test.go relies on. Asserts on the is_closed flag round-
+// trip rather than specific status names so the test is portable across
+// Redmine default-data variations.
 func TestMetadata_StatusesList(t *testing.T) {
 	requireE2E(t)
 	r := newCLIRunner(t, e2eBaseURL(), e2eAPIKey())
@@ -44,14 +49,10 @@ func TestMetadata_StatusesList(t *testing.T) {
 		t.Fatal("statuses list returned no statuses")
 	}
 
-	hasClosed := false
-	hasOpen := false
+	var hasClosed, hasOpen bool
 	for _, s := range statuses {
-		if s.ID == 0 {
-			t.Errorf("status has zero ID: %+v", s)
-		}
-		if s.Name == "" {
-			t.Errorf("status has empty name: %+v", s)
+		if s.ID == 0 || s.Name == "" {
+			t.Errorf("status entry missing id or name: %+v", s)
 		}
 		if s.IsClosed {
 			hasClosed = true
